Rewrite unit test with window query, and add more test
